pyproject.toml
[project]
name = "ort"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "django-nested-admin>=4.1.6",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]
